downloaded input system
i set up the input actions asset and played around a bit in a temp script to get a feel for how the events are called.
hmm i think for now the scripting is okay. i should focus more on making more concept art and maybe animations for the character now.